+++ calculate/commands/real/add.h
@@ -3,8 +3,8 @@
 
 #include "command.h"
 __global__ void f_add(const double* inA, const double* inB, double* out, size_t size) {
-    size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
-    int stride = gridDim.x * blockDim.x;
+    uint idx = blockIdx.x * blockDim.x + threadIdx.x;
+    uint stride = gridDim.x * blockDim.x;
     for (size_t i = idx; i < size; i += stride) {
         out[i] = inA[i] + inB[i];
     }
@@ -31,8 +31,8 @@ public:
 
 
 __global__ void f_add_c(const double* in, double ref, double* out, size_t size) {
-    size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
-    int stride = gridDim.x * blockDim.x;
+    uint idx = blockIdx.x * blockDim.x + threadIdx.x;
+    uint stride = gridDim.x * blockDim.x;
     for (size_t i = idx; i < size; i += stride) {
         out[i] = in[i] + ref;
     }
