+++ calculate/commands/real/div.h
@@ -4,7 +4,7 @@
 #include "command.h"
 #include <cmath>
 __global__ void f_div(const double* inA, const double* inB, double* out, size_t size) {
-    size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+    uint idx = blockIdx.x * blockDim.x + threadIdx.x;
     uint stride = gridDim.x * blockDim.x;
     for (size_t i = idx; i < size; i += stride) {
         out[i] = inA[i] / inB[i];
@@ -32,7 +32,7 @@ public:
 
 
 __global__ void f_div_c(const double* in, double ref, double* out, size_t size) {
-    size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+    uint idx = blockIdx.x * blockDim.x + threadIdx.x;
     uint stride = gridDim.x * blockDim.x;
     for (size_t i = idx; i < size; i += stride) {
         out[i] = ref / in[i];
