+++ calculate/commands/real/sub.h
@@ -3,7 +3,7 @@
 
 #include "command.h"
 __global__ void f_sub(const double* inA, const double* inB, double* out, size_t size) {
-    size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+    uint idx = blockIdx.x * blockDim.x + threadIdx.x;
     uint stride = gridDim.x * blockDim.x;
     for (size_t i = idx; i < size; i += stride) {
         out[i] = inA[i] - inB[i];
@@ -30,7 +30,7 @@ public:
 
 
 __global__ void f_sub_c(const double* in, double ref, double* out, size_t size) {
-    size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+    uint idx = blockIdx.x * blockDim.x + threadIdx.x;
     uint stride = gridDim.x * blockDim.x;
     for (size_t i = idx; i < size; i += stride) {
         out[i] = in[i] - ref;
